Rename config module to util.config.
Provides a natural place to put other types of utilities.
